Story #11756 : Fix support for Windows ZIPs with implicit parent folder entries

Windows explorer generates ZIPs while skipping "folder" entries as they are considered "implicit" (i.e while most linux tools generate "A/B/C.txt" as a "A/" folder, "A/B/" folder and a "A/B/C.txt" file, windows only creates a single "A/B/C.txt" entry)

Also fixes upload of ZIPs with empty files

See merge request vitam/vitam!9903

+++ sources/collect/collect-internal/collect-internal-core/src/main/java/fr/gouv/vitam/collect/internal/core/service/FluxService.java
@@ -32,10 +32,10 @@ import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.fasterxml.jackson.databind.node.TextNode;
 import com.google.common.collect.Iterators;
 import fr.gouv.culture.archivesdefrance.seda.v2.LevelType;
+import fr.gouv.vitam.collect.common.exception.CollectInternalClientInvalidRequestException;
 import fr.gouv.vitam.collect.common.exception.CollectInternalException;
 import fr.gouv.vitam.collect.common.exception.CsvParseInternalException;
 import fr.gouv.vitam.collect.internal.core.common.ProjectModel;
-import fr.gouv.vitam.collect.internal.core.common.TransactionModel;
 import fr.gouv.vitam.collect.internal.core.helpers.CsvHelper;
 import fr.gouv.vitam.collect.internal.core.helpers.MetadataHelper;
 import fr.gouv.vitam.collect.internal.core.repository.MetadataRepository;
@@ -138,6 +138,9 @@ public class FluxService {
             int maxLevel = -1;
             while ((entry = archiveInputStream.getNextEntry()) != null) {
                 if (archiveInputStream.canReadEntryData(entry)) {
+
+                    checkNonEmptyBinary(entry);
+
                     String path = FilenameUtils.normalize(entry.getName());
                     if (!FilenameUtils.equals(entry.getName(), path)) {
                         throw new IllegalStateException("path " + path + " is not canonical");
@@ -152,8 +155,8 @@ public class FluxService {
                             METADATA_CSV_FILE);
                         isExtraMetadataExist = true;
                     } else {
-                        maxLevel = createMetadata(transactionId, entry, entryInputStream, maxLevel, unitIds, path,
-                            projectModel.getUnitUp() != null);
+                        maxLevel = createMetadata(transactionId, path, entryInputStream, entry.isDirectory(), maxLevel,
+                            unitIds, projectModel.getUnitUp() != null);
                     }
                     isEmpty = false;
                 }
@@ -180,8 +183,6 @@ public class FluxService {
 
             bulkWriteObjectGroups(transactionId);
 
-            cleanTemporaryFiles(maxLevel, transactionId);
-
             if (isExtraMetadataExist) {
                 File metadataFile = PropertiesUtils.fileFromTmpFolder(
                     METADATA + "_" + transactionId + VitamConstants.JSONL_EXTENSION);
@@ -196,16 +197,30 @@ public class FluxService {
             throw new CollectInternalException("An error occurs when try to upload the ZIP: {}");
         } catch (InvalidParseOperationException | CsvParseInternalException e) {
             throw new CollectInternalException(e.getMessage(), e);
+        } finally {
+            cleanTemporaryFiles(transactionId);
         }
     }
 
-    private void cleanTemporaryFiles(int maxLevel, String transactionId) {
+    private void checkNonEmptyBinary(ArchiveEntry entry) throws CollectInternalClientInvalidRequestException {
+        if (!entry.isDirectory() && entry.getSize() == 0L) {
+            throw new CollectInternalClientInvalidRequestException(
+                "Cannot upload empty file '" + entry.getName() + "'");
+        }
+    }
+
+    private void cleanTemporaryFiles(String transactionId) {
         File ogFile = PropertiesUtils.fileFromTmpFolder(
             MetadataType.OBJECTGROUP.getName() + "_" + transactionId + VitamConstants.JSONL_EXTENSION);
         FileUtils.deleteQuietly(ogFile);
-        for (int level = 0; level < maxLevel; level++) {
+        for (int level = 0; ; level++) {
             File file = PropertiesUtils.fileFromTmpFolder(
                 MetadataType.UNIT.getName() + "_" + level + "_" + transactionId + VitamConstants.JSONL_EXTENSION);
+
+            if (!file.exists()) {
+                // No more files...
+                return;
+            }
             FileUtils.deleteQuietly(file);
         }
     }
@@ -233,25 +248,27 @@ public class FluxService {
         }
     }
 
-    private int createMetadata(String transactionId, ArchiveEntry entry,
-        ArchiveEntryInputStream entryInputStream, int maxLevel, Map<String, String> unitIds, String path,
-        boolean isAttachmentAuExist) throws IOException, CollectInternalException, InvalidParseOperationException {
-        LevelType descriptionLevel = (entry.isDirectory()) ? LevelType.RECORD_GRP : LevelType.ITEM;
-
-        String parent = FilenameUtils.getPathNoEndSeparator(path);
+    private int createMetadata(String transactionId, String path, InputStream entryInputStream, boolean isDirectory,
+        int maxLevel, Map<String, String> unitIds, boolean isAttachmentAuExist)
+        throws IOException, CollectInternalException, InvalidParseOperationException {
+        LevelType descriptionLevel = isDirectory ? LevelType.RECORD_GRP : LevelType.ITEM;
+        String parentPath = FilenameUtils.getPathNoEndSeparator(path);
 
         String parentUnit;
-        if (Strings.isNullOrEmpty(parent)) {
+        if (Strings.isNullOrEmpty(parentPath)) {
             if (isAttachmentAuExist) {
                 parentUnit = unitIds.get(STATIC_ATTACHMENT);
             } else {
                 parentUnit = null;
             }
         } else {
-            parentUnit = unitIds.get(parent);
+            parentUnit = unitIds.get(parentPath);
             if (parentUnit == null) {
-                throw new IllegalStateException("Malformed zip file : cannot import tree");
+                LOGGER.debug("Creating implicit parent folder '{}'", parentPath);
+                createMetadata(transactionId, parentPath, null, true, maxLevel, unitIds, isAttachmentAuExist);
             }
+
+            parentUnit = unitIds.get(parentPath);
         }
         String fileName = FilenameUtils.getName(path);
 
@@ -259,7 +276,7 @@ public class FluxService {
             MetadataHelper.createUnit(transactionId, descriptionLevel, fileName, parentUnit);
 
         unitIds.put(path, unit.getId());
-        if (!entry.isDirectory()) {
+        if (!isDirectory) {
             String extension = FilenameUtils.getExtension(fileName).toLowerCase();
             String objectId = GUIDFactory.newGUID().getId();
             String newFilename = (Strings.isNullOrEmpty(extension)) ? objectId : objectId + "." + extension;
